[bug] Fix build.py Powershell part. (#8291)
Issue: #

### Brief Summary

"-Interactive" seems not to be a valid argument for the current
Powershell. On my Windows 11 machine, invoke `python build.py --shell`
gives the following error:


![image](https://github.com/taichi-dev/taichi/assets/2747993/fd58e452-30fd-4d0f-a1fd-145c1fa3f548)

This PR removes the "-Interactive" flag, and add a "ExecutionPolicy"
flag which is required for running the remaining part of the `build.py`
script.