[autodiff] Fix the type of cmp statements in autodiff (#8452)
Issue: fixes #8444

The return type of cmp statements of tensors should be tensors of u1
instead of tensors of i32.

Sometimes the CFG detects that an AdStackLoadTopStmt and an
AdStackLoadTopAdjStmt loading the same address. I don't know if this
should happen, but it stops raising error if I don't eliminate the
latter statement.