[amdgpu] Calculate mcpu_ and compute_capability_ properly and with ROCm 6 compat (#8667)
Issue: #6434 (part of)

### Brief Summary

In the AMDGPU backend, we calculate `mcpu_` directly based on
`compute_capability_` using integer addition. However, this can be
problematic as the `mcpu_` name is partially hex-based, e.g.
`compute_capability_ = 910` corresponds to `mcpu_ = "gfx90a"`. The
proper and
[recommended](https://github.com/ROCm/hip/blob/80681169ae20de8f8025b4fad799204c3ae7de50/include/hip/hip_runtime_api.h#L129)
way to get `mcpu_` is using field `gcnArchName`. Similarly,
`compute_capability_` should be calculated using fields `major` and
`minor` instead of `gcnArch`.

Additionally, there are complications in ROCm 6 regarding calling
`hipGetDeviceProperties` by looking up its ABI symbol in
`libamdhip64.so`. In ROCm 6, the ABI symbol `hipGetDeviceProperties`
(likely) incorrectly maps to the ROCm 5 version of
`hipGetDeviceProperties` which is not ABI-compatible. To handle this, we
first treat `hipGetDeviceProperties` as ROCm 5 version, and if the
values we get don't make sense we then treat it as the ROCm 6 version.

### Walkthrough

copilot:walkthrough

---------

Signed-off-by: Gavin Zhao <git@gzgz.dev>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>