[GUI] Support ndarray input (#8661)
Currently ggui only accept taichi.Field as some of its input like
centers, colors.
I think the reason for this might be that some of these attributes are
optional so that we must use taichi.template() to compile accordingly.
However, I found no way to achieve the same with ndarray. So I split the
writing process into mutiple kernels when it comes to ndarray.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>